Rename queries to import-queries
This means we can keep our import and clean queries distinct.
